+++ CHANGELOG.rst
@@ -22,6 +22,8 @@ not released yet
 * FIX display search results that are all-day events
 * in ikhal, update the date-titles on date change
 * FIX printing a new event's path if [default] print_new = path
+* FIX width of calendar in `khal calendar` was off by two if locale.weeknumbers
+  was set to "right"
 
 0.9.1
 =====