+++ CHANGELOG.rst
@@ -7,6 +7,12 @@ Package maintainers and users who have to manually update their installation
 may want to subscribe to `GitHub's tag feed
 <https://github.com/geier/khal/tags.atom>`_.
 
+0.9.3
+=====
+not released yet
+
+* FIX `configure` would crash if neither "import config from vdirsyncer" nor
+  "create locale vdir" was selected
 
 0.9.2
 =====